Fixes for departmental PC

+++ prod_quan_nn.cpp
@@ -29,7 +29,7 @@ namespace bdap {
     ProdQuanNN::initialize_method()
     {
         //std::cout << "Construct auxiliary structures here" << std::endl;
-        for (int i = 0; i < this->npartitions(); i++){
+        for (int i = 0; i < (int) this->npartitions(); i++){
             this->clusterDistance.push_back( std::vector<float>(this->nclusters(i)));
         }
     }
@@ -42,11 +42,11 @@ namespace bdap {
                 pydata<float>& out_distance) const
     {
         std::vector<std::vector<float>> clusterDistance;
-        for (int i = 0; i < this->npartitions(); i++){
+        for (int i = 0; i < (int) this->npartitions(); i++){
             clusterDistance.push_back( std::vector<float>(this->nclusters(i)));
         }
 
-        for (int i = 0; i < examples.nrows; i++){
+        for (int i = 0; i < (int) examples.nrows; i++){
             this->compute_nearest_single(examples, nneighbors, out_index, out_distance, i, clusterDistance);
         }
 
@@ -57,8 +57,8 @@ namespace bdap {
     void ProdQuanNN::compute_nearest_single(const pydata<float>& examples, int nneighbors, pydata<int>& out_index, pydata<float>& out_distance, int index,
     std::vector<std::vector<float>>& clusterDistance) const{
         // Initialize the distance to each cluster
-        for (int p = 0; p < this->npartitions(); p++){
-            for (int c = 0; c < this->nclusters(p); c++){
+        for (int p = 0; p < (int) this->npartitions(); p++){
+            for (int c = 0; c < (int) this->nclusters(p); c++){
                 clusterDistance[p][c] = this->distance_to_cluster(examples, index, p, c);
             }
         }
@@ -66,9 +66,9 @@ namespace bdap {
         // KNN
         std::map<float, int> queue;
         float max_distance = 0;
-        for (int i=0; i < this->ntrain_examples(); i++){
+        for (int i=0; i < (int) this->ntrain_examples(); i++){
             float distance = this->distance_to_label(i, clusterDistance);
-            if (queue.size() < nneighbors){
+            if (queue.size() < (size_t) nneighbors){
                 queue[distance] = i;
                 max_distance = std::max(max_distance, distance);
             } else if (distance < max_distance){
@@ -95,7 +95,7 @@ namespace bdap {
         const float* ptr = this->centroid(partition, cluster);
         int begin_feature = this->feat_begin(partition);
         float distance = 0;
-        for (int i = begin_feature; i <  this->feat_end(partition); i++){
+        for (int i = begin_feature; i < (int) this->feat_end(partition); i++){
             distance += std::pow(*(ptr + (i - begin_feature)) - examples.get_elem(example_index, i), 2);
         }
         distance = std::sqrt(distance);
@@ -104,7 +104,7 @@ namespace bdap {
 
     float ProdQuanNN::distance_to_label(size_t label_index, std::vector<std::vector<float>>& clusterDistance) const{
         float distance = 0;
-        for (int p = 0; p < this->npartitions(); p++){
+        for (int p = 0; p < (int) this->npartitions(); p++){
             
             int cluster = *(this->labels(p)+ label_index);
